Remove FrontstageDef.getWidgetCurrentState
